\name{list}
\alias{list}
\docType{data}
\title{
AAindex database
}
\description{
The modified AAindex database with 532 descriptors for the twenty amino acids.
}
\usage{data(list)}
\format{
}
\details{
1	alpha-CH_chemical_shifts_(Andersen_et_al.,_1992)

2	Hydrophobicity_index_(Argos_et_al.,_1982)

3	Signal_sequence_helical_potential_(Argos_et_al.,_1982)

4	Membrane-buried_preference_parameters_(Argos_et_al.,_1982)

5	Conformational_parameter_of_inner_helix_(Beghin-Dirkx,_1975)

6	Conformational_parameter_of_beta-structure_(Beghin-Dirkx,_1975)

7	Conformational_parameter_of_beta-turn_(Beghin-Dirkx,_1975)

8	Average_flexibility_indices_(Bhaskaran-Ponnuswamy,_1988)

9	Residue_volume_(Bigelow,_1967)

10	Information_value_for_accessibility;_average_fraction_35%_(Biou_et_al.,_1988)

11	Information_value_for_accessibility;_average_fraction_23%_(Biou_et_al.,_1988)

12	Retention_coefficient_in_TFA_(Browne_et_al.,_1982)

13	Retention_coefficient_in_HFBA_(Browne_et_al.,_1982)

14	Transfer_free_energy_to_surface_(Bull-Breese,_1974)

15	Apparent_partial_specific_volume_(Bull-Breese,_1974)

16	alpha-NH_chemical_shifts_(Bundi-Wuthrich,_1979)

17	alpha-CH_chemical_shifts_(Bundi-Wuthrich,_1979)

18	Spin-spin_coupling_constants_3JHalpha-NH_(Bundi-Wuthrich,_1979)

19	Normalized_frequency_of_alpha-helix_(Burgess_et_al.,_1974)

20	Normalized_frequency_of_extended_structure_(Burgess_et_al.,_1974)

21	Steric_parameter_(Charton,_1981)

22	Polarizability_parameter_(Charton-Charton,_1982)

23	Free_energy_of_solution_in_water,_kcal/mole_(Charton-Charton,_1982)

24	The_Chou-Fasman_parameter_of_the_coil_conformation_(Charton-Charton,_1983)

25	A_parameter_defined_from_the_residuals_obtained_from_the_best_correlation_of_the_Chou-Fasman_parameter_of_beta-sheet_(Charton-Charton,_1983)

26	The_number_of_atoms_in_the_side_chain_labelled_1+1_(Charton-Charton,_1983)

27	The_number_of_atoms_in_the_side_chain_labelled_2+1_(Charton-Charton,_1983)

28	The_number_of_atoms_in_the_side_chain_labelled_3+1_(Charton-Charton,_1983)

29	The_number_of_bonds_in_the_longest_chain_(Charton-Charton,_1983)

30	A_parameter_of_charge_transfer_capability_(Charton-Charton,_1983)

31	A_parameter_of_charge_transfer_donor_capability_(Charton-Charton,_1983)

32	Average_volume_of_buried_residue_(Chothia,_1975)

33	Residue_accessible_surface_area_in_tripeptide_(Chothia,_1976)

34	Residue_accessible_surface_area_in_folded_protein_(Chothia,_1976)

35	Proportion_of_residues_95%_buried_(Chothia,_1976)

36	Proportion_of_residues_100%_buried_(Chothia,_1976)

37	Normalized_frequency_of_beta-turn_(Chou-Fasman,_1978a)

38	Normalized_frequency_of_alpha-helix_(Chou-Fasman,_1978b)

39	Normalized_frequency_of_beta-sheet_(Chou-Fasman,_1978b)

40	Normalized_frequency_of_beta-turn_(Chou-Fasman,_1978b)

41	Normalized_frequency_of_N-terminal_helix_(Chou-Fasman,_1978b)

42	Normalized_frequency_of_C-terminal_helix_(Chou-Fasman,_1978b)

43	Normalized_frequency_of_N-terminal_non_helical_region_(Chou-Fasman,_1978b)

44	Normalized_frequency_of_C-terminal_non_helical_region_(Chou-Fasman,_1978b)

45	Normalized_frequency_of_N-terminal_beta-sheet_(Chou-Fasman,_1978b)

46	Normalized_frequency_of_C-terminal_beta-sheet_(Chou-Fasman,_1978b)

47	Normalized_frequency_of_N-terminal_non_beta_region_(Chou-Fasman,_1978b)

48	Normalized_frequency_of_C-terminal_non_beta_region_(Chou-Fasman,_1978b)

49	Frequency_of_the_1st_residue_in_turn_(Chou-Fasman,_1978b)

50	Frequency_of_the_2nd_residue_in_turn_(Chou-Fasman,_1978b)

51	Frequency_of_the_3rd_residue_in_turn_(Chou-Fasman,_1978b)

52	Frequency_of_the_4th_residue_in_turn_(Chou-Fasman,_1978b)

53	Normalized_frequency_of_the_2nd_and_3rd_residues_in_turn_(Chou-Fasman,_1978b)

54	Normalized_hydrophobicity_scales_for_alpha-proteins_(Cid_et_al.,_1992)

55	Normalized_hydrophobicity_scales_for_beta-proteins_(Cid_et_al.,_1992)

56	Normalized_hydrophobicity_scales_for_alpha+beta-proteins_(Cid_et_al.,_1992)

57	Normalized_hydrophobicity_scales_for_alpha/beta-proteins_(Cid_et_al.,_1992)

58	Normalized_average_hydrophobicity_scales_(Cid_et_al.,_1992)

59	Partial_specific_volume_(Cohn-Edsall,_1943)

60	Normalized_frequency_of_middle_helix_(Crawford_et_al.,_1973)

61	Normalized_frequency_of_beta-sheet_(Crawford_et_al.,_1973)

62	Normalized_frequency_of_turn_(Crawford_et_al.,_1973)

63	Size_(Dawson,_1972)

64	Amino_acid_composition_(Dayhoff_et_al.,_1978a)

65	Relative_mutability_(Dayhoff_et_al.,_1978b)

66	Membrane_preference_for_cytochrome_b:_MPH89_(Degli_Esposti_et_al.,_1990)

67	Average_membrane_preference:_AMP07

68	Consensus_normalized_hydrophobicity_scale_(Eisenberg,_1984)

69	Solvation_free_energy_(Eisenberg-McLachlan,_1986)

70	Atom-based_hydrophobic_moment_(Eisenberg-McLachlan,_1986)

71	Direction_of_hydrophobic_moment_(Eisenberg-McLachlan,_1986)

72	Molecular_weight_(Fasman,_1976)

73	Melting_point_(Fasman,_1976)

74	Optical_rotation_(Fasman,_1976)

75	pK-N_(Fasman,_1976)

76	pK-C_(Fasman,_1976)

77	Hydrophobic_parameter_pi_(Fauchere-Pliska,_1983)

78	Graph_shape_index_(Fauchere_et_al.,_1988)

79	Smoothed_upsilon_steric_parameter_(Fauchere_et_al.,_1988)

80	Normalized_van_der_Waals_volume_(Fauchere_et_al.,_1988)

81	STERIMOL_length_of_the_side_chain_(Fauchere_et_al.,_1988)

82	STERIMOL_minimum_width_of_the_side_chain_(Fauchere_et_al.,_1988)

83	STERIMOL_maximum_width_of_the_side_chain_(Fauchere_et_al.,_1988)

84	N.m.r._chemical_shift_of_alpha-carbon_(Fauchere_et_al.,_1988)

85	Localized_electrical_effect_(Fauchere_et_al.,_1988)

86	Number_of_hydrogen_bond_donors_(Fauchere_et_al.,_1988)

87	Number_of_full_nonbonding_orbitals_(Fauchere_et_al.,_1988)

88	Positive_charge_(Fauchere_et_al.,_1988)

89	Negative_charge_(Fauchere_et_al.,_1988)

90	pK-a(RCOOH)_(Fauchere_et_al.,_1988)

91	Helix-coil_equilibrium_constant_(Finkelstein-Ptitsyn,_1977)

92	Helix_initiation_parameter_at_posision_i-1_(Finkelstein_et_al.,_1991)

93	Helix_initiation_parameter_at_posision_i,i+1,i+2_(Finkelstein_et_al.,_1991)

94	Helix_termination_parameter_at_posision_j-2,j-1,j_(Finkelstein_et_al.,_1991)

95	Helix_termination_parameter_at_posision_j+1_(Finkelstein_et_al.,_1991)

96	Partition_coefficient_(Garel_et_al.,_1973)

97	Alpha-helix_indices_(Geisow-Roberts,_1980)

98	Alpha-helix_indices_for_alpha-proteins_(Geisow-Roberts,_1980)

99	Alpha-helix_indices_for_beta-proteins_(Geisow-Roberts,_1980)

100	Alpha-helix_indices_for_alpha/beta-proteins_(Geisow-Roberts,_1980)

101	Beta-strand_indices_(Geisow-Roberts,_1980)

102	Beta-strand_indices_for_beta-proteins_(Geisow-Roberts,_1980)

103	Beta-strand_indices_for_alpha/beta-proteins_(Geisow-Roberts,_1980)

104	Aperiodic_indices_(Geisow-Roberts,_1980)

105	Aperiodic_indices_for_alpha-proteins_(Geisow-Roberts,_1980)

106	Aperiodic_indices_for_beta-proteins_(Geisow-Roberts,_1980)

107	Aperiodic_indices_for_alpha/beta-proteins_(Geisow-Roberts,_1980)

108	Hydrophobicity_factor_(Goldsack-Chalifoux,_1973)

109	Residue_volume_(Goldsack-Chalifoux,_1973)

110	Composition_(Grantham,_1974)

111	Polarity_(Grantham,_1974)

112	Volume_(Grantham,_1974)

113	Partition_energy_(Guy,_1985)

114	Hydration_number_(Hopfinger,_1971),_Cited_by_Charton-Charton_(1982)

115	Hydrophilicity_value_(Hopp-Woods,_1981)

116	Heat_capacity_(Hutchens,_1970)

117	Absolute_entropy_(Hutchens,_1970)

118	Entropy_of_formation_(Hutchens,_1970)

119	Normalized_relative_frequency_of_alpha-helix_(Isogai_et_al.,_1980)

120	Normalized_relative_frequency_of_extended_structure_(Isogai_et_al.,_1980)

121	Normalized_relative_frequency_of_bend_(Isogai_et_al.,_1980)

122	Normalized_relative_frequency_of_bend_R_(Isogai_et_al.,_1980)

123	Normalized_relative_frequency_of_bend_S_(Isogai_et_al.,_1980)

124	Normalized_relative_frequency_of_helix_end_(Isogai_et_al.,_1980)

125	Normalized_relative_frequency_of_double_bend_(Isogai_et_al.,_1980)

126	Normalized_relative_frequency_of_coil_(Isogai_et_al.,_1980)

127	Average_accessible_surface_area_(Janin_et_al.,_1978)

128	Percentage_of_buried_residues_(Janin_et_al.,_1978)

129	Percentage_of_exposed_residues_(Janin_et_al.,_1978)

130	Ratio_of_buried_and_accessible_molar_fractions_(Janin,_1979)

131	Transfer_free_energy_(Janin,_1979)

132	Hydrophobicity_(Jones,_1975)

133	pK_(-COOH)_(Jones,_1975)

134	Relative_frequency_of_occurrence_(Jones_et_al.,_1992)

135	Relative_mutability_(Jones_et_al.,_1992)

136	Amino_acid_distribution_(Jukes_et_al.,_1975)

137	Sequence_frequency_(Jungck,_1978)

138	Average_relative_probability_of_helix_(Kanehisa-Tsong,_1980)

139	Average_relative_probability_of_beta-sheet_(Kanehisa-Tsong,_1980)

140	Average_relative_probability_of_inner_helix_(Kanehisa-Tsong,_1980)

141	Average_relative_probability_of_inner_beta-sheet_(Kanehisa-Tsong,_1980)

142	Flexibility_parameter_for_no_rigid_neighbors_(Karplus-Schulz,_1985)

143	Flexibility_parameter_for_one_rigid_neighbor_(Karplus-Schulz,_1985)

144	Flexibility_parameter_for_two_rigid_neighbors_(Karplus-Schulz,_1985)

145	The_Kerr-constant_increments_(Khanarian-Moore,_1980)

146	Net_charge_(Klein_et_al.,_1984)

147	Side_chain_interaction_parameter_(Krigbaum-Rubin,_1971)

148	Side_chain_interaction_parameter_(Krigbaum-Komoriya,_1979)

149	Fraction_of_site_occupied_by_water_(Krigbaum-Komoriya,_1979)

150	Side_chain_volume_(Krigbaum-Komoriya,_1979)

151	Hydropathy_index_(Kyte-Doolittle,_1982)

152	Transfer_free_energy,_CHP/water_(Lawson_et_al.,_1984)

153	Hydrophobic_parameter_(Levitt,_1976)

154	Distance_between_C-alpha_and_centroid_of_side_chain_(Levitt,_1976)

155	Side_chain_angle_theta(AAR)_(Levitt,_1976)

156	Side_chain_torsion_angle_phi(AAAR)_(Levitt,_1976)

157	Radius_of_gyration_of_side_chain_(Levitt,_1976)

158	van_der_Waals_parameter_R0_(Levitt,_1976)

159	van_der_Waals_parameter_epsilon_(Levitt,_1976)

160	Normalized_frequency_of_alpha-helix,_with_weights_(Levitt,_1978)

161	Normalized_frequency_of_beta-sheet,_with_weights_(Levitt,_1978)

162	Normalized_frequency_of_reverse_turn,_with_weights_(Levitt,_1978)

163	Normalized_frequency_of_alpha-helix,_unweighted_(Levitt,_1978)

164	Normalized_frequency_of_beta-sheet,_unweighted_(Levitt,_1978)

165	Normalized_frequency_of_reverse_turn,_unweighted_(Levitt,_1978)

166	Frequency_of_occurrence_in_beta-bends_(Lewis_et_al.,_1971)

167	Conformational_preference_for_all_beta-strands_(Lifson-Sander,_1979)

168	Conformational_preference_for_parallel_beta-strands_(Lifson-Sander,_1979)

169	Conformational_preference_for_antiparallel_beta-strands_(Lifson-Sander,_1979)

170	Average_surrounding_hydrophobicity_(Manavalan-Ponnuswamy,_1978)

171	Normalized_frequency_of_alpha-helix_(Maxfield-Scheraga,_1976)

172	Normalized_frequency_of_extended_structure_(Maxfield-Scheraga,_1976)

173	Normalized_frequency_of_zeta_R_(Maxfield-Scheraga,_1976)

174	Normalized_frequency_of_left-handed_alpha-helix_(Maxfield-Scheraga,_1976)

175	Normalized_frequency_of_zeta_L_(Maxfield-Scheraga,_1976)

176	Normalized_frequency_of_alpha_region_(Maxfield-Scheraga,_1976)

177	Refractivity_(McMeekin_et_al.,_1964),_Cited_by_Jones_(1975)

178	Retention_coefficient_in_HPLC,_pH7.4_(Meek,_1980)

179	Retention_coefficient_in_HPLC,_pH2.1_(Meek,_1980)

180	Retention_coefficient_in_NaClO4_(Meek-Rossetti,_1981)

181	Retention_coefficient_in_NaH2PO4_(Meek-Rossetti,_1981)

182	Average_reduced_distance_for_C-alpha_(Meirovitch_et_al.,_1980)

183	Average_reduced_distance_for_side_chain_(Meirovitch_et_al.,_1980)

184	Average_side_chain_orientation_angle_(Meirovitch_et_al.,_1980)

185	Effective_partition_energy_(Miyazawa-Jernigan,_1985)

186	Normalized_frequency_of_alpha-helix_(Nagano,_1973)

187	Normalized_frequency_of_bata-structure_(Nagano,_1973)

188	Normalized_frequency_of_coil_(Nagano,_1973)

189	AA_composition_of_total_proteins_(Nakashima_et_al.,_1990)

190	SD_of_AA_composition_of_total_proteins_(Nakashima_et_al.,_1990)

191	AA_composition_of_mt-proteins_(Nakashima_et_al.,_1990)

192	Normalized_composition_of_mt-proteins_(Nakashima_et_al.,_1990)

193	AA_composition_of_mt-proteins_from_animal_(Nakashima_et_al.,_1990)

194	Normalized_composition_from_animal_(Nakashima_et_al.,_1990)

195	AA_composition_of_mt-proteins_from_fungi_and_plant_(Nakashima_et_al.,_1990)

196	Normalized_composition_from_fungi_and_plant_(Nakashima_et_al.,_1990)

197	AA_composition_of_membrane_proteins_(Nakashima_et_al.,_1990)

198	Normalized_composition_of_membrane_proteins_(Nakashima_et_al.,_1990)

199	Transmembrane_regions_of_non-mt-proteins_(Nakashima_et_al.,_1990)

200	Transmembrane_regions_of_mt-proteins_(Nakashima_et_al.,_1990)

201	Ratio_of_average_and_computed_composition_(Nakashima_et_al.,_1990)

202	AA_composition_of_CYT_of_single-spanning_proteins_(Nakashima-Nishikawa,_1992)

203	AA_composition_of_CYT2_of_single-spanning_proteins_(Nakashima-Nishikawa,_1992)

204	AA_composition_of_EXT_of_single-spanning_proteins_(Nakashima-Nishikawa,_1992)

205	AA_composition_of_EXT2_of_single-spanning_proteins_(Nakashima-Nishikawa,_1992)

206	AA_composition_of_MEM_of_single-spanning_proteins_(Nakashima-Nishikawa,_1992)

207	AA_composition_of_CYT_of_multi-spanning_proteins_(Nakashima-Nishikawa,_1992)

208	AA_composition_of_EXT_of_multi-spanning_proteins_(Nakashima-Nishikawa,_1992)

209	AA_composition_of_MEM_of_multi-spanning_proteins_(Nakashima-Nishikawa,_1992)

210	8_A_contact_number_(Nishikawa-Ooi,_1980)

211	14

212	Transfer_energy,_organic_solvent/water_(Nozaki-Tanford,_1971)

213	Average_non-bonded_energy_per_atom_(Oobatake-Ooi,_1977)

214	Short_and_medium_range_non-bonded_energy_per_atom_(Oobatake-Ooi,_1977)

215	Long_range_non-bonded_energy_per_atom_(Oobatake-Ooi,_1977)

216	Average_non-bonded_energy_per_residue_(Oobatake-Ooi,_1977)

217	Short_and_medium_range_non-bonded_energy_per_residue_(Oobatake-Ooi,_1977)

218	Optimized_beta-structure-coil_equilibrium_constant_(Oobatake_et_al.,_1985)

219	Optimized_propensity_to_form_reverse_turn_(Oobatake_et_al.,_1985)

220	Optimized_transfer_energy_parameter_(Oobatake_et_al.,_1985)

221	Optimized_average_non-bonded_energy_per_atom_(Oobatake_et_al.,_1985)

222	Optimized_side_chain_interaction_parameter_(Oobatake_et_al.,_1985)

223	Normalized_frequency_of_alpha-helix_from_LG_(Palau_et_al.,_1981)

224	Normalized_frequency_of_alpha-helix_from_CF_(Palau_et_al.,_1981)

225	Normalized_frequency_of_beta-sheet_from_LG_(Palau_et_al.,_1981)

226	Normalized_frequency_of_beta-sheet_from_CF_(Palau_et_al.,_1981)

227	Normalized_frequency_of_turn_from_LG_(Palau_et_al.,_1981)

228	Normalized_frequency_of_turn_from_CF_(Palau_et_al.,_1981)

229	Normalized_frequency_of_alpha-helix_in_all-alpha_class_(Palau_et_al.,_1981)

230	Normalized_frequency_of_alpha-helix_in_alpha+beta_class_(Palau_et_al.,_1981)

231	Normalized_frequency_of_alpha-helix_in_alpha/beta_class_(Palau_et_al.,_1981)

232	Normalized_frequency_of_beta-sheet_in_all-beta_class_(Palau_et_al.,_1981)

233	Normalized_frequency_of_beta-sheet_in_alpha+beta_class_(Palau_et_al.,_1981)

234	Normalized_frequency_of_beta-sheet_in_alpha/beta_class_(Palau_et_al.,_1981)

235	Normalized_frequency_of_turn_in_all-alpha_class_(Palau_et_al.,_1981)

236	Normalized_frequency_of_turn_in_all-beta_class_(Palau_et_al.,_1981)

237	Normalized_frequency_of_turn_in_alpha+beta_class_(Palau_et_al.,_1981)

238	Normalized_frequency_of_turn_in_alpha/beta_class_(Palau_et_al.,_1981)

239	HPLC_parameter_(Parker_et_al.,_1986)

240	Partition_coefficient_(Pliska_et_al.,_1981)

241	Surrounding_hydrophobicity_in_folded_form_(Ponnuswamy_et_al.,_1980)

242	Average_gain_in_surrounding_hydrophobicity_(Ponnuswamy_et_al.,_1980)

243	Average_gain_ratio_in_surrounding_hydrophobicity_(Ponnuswamy_et_al.,_1980)

244	Surrounding_hydrophobicity_in_alpha-helix_(Ponnuswamy_et_al.,_1980)

245	Surrounding_hydrophobicity_in_beta-sheet_(Ponnuswamy_et_al.,_1980)

246	Surrounding_hydrophobicity_in_turn_(Ponnuswamy_et_al.,_1980)

247	Accessibility_reduction_ratio_(Ponnuswamy_et_al.,_1980)

248	Average_number_of_surrounding_residues_(Ponnuswamy_et_al.,_1980)

249	Intercept_in_regression_analysis_(Prabhakaran-Ponnuswamy,_1982)

250	Slope_in_regression_analysis_x_1.0E1_(Prabhakaran-Ponnuswamy,_1982)

251	Correlation_coefficient_in_regression_analysis_(Prabhakaran-Ponnuswamy,_1982)

252	Hydrophobicity_(Prabhakaran,_1990)

253	Relative_frequency_in_alpha-helix_(Prabhakaran,_1990)

254	Relative_frequency_in_beta-sheet_(Prabhakaran,_1990)

255	Relative_frequency_in_reverse-turn_(Prabhakaran,_1990)

256	Helix-coil_equilibrium_constant_(Ptitsyn-Finkelstein,_1983)

257	Beta-coil_equilibrium_constant_(Ptitsyn-Finkelstein,_1983)

258	Weights_for_alpha-helix_at_the_window_position_of_-6_(Qian-Sejnowski,_1988)

259	Weights_for_alpha-helix_at_the_window_position_of_-5_(Qian-Sejnowski,_1988)

260	Weights_for_alpha-helix_at_the_window_position_of_-4_(Qian-Sejnowski,_1988)

261	Weights_for_alpha-helix_at_the_window_position_of_-3_(Qian-Sejnowski,_1988)

262	Weights_for_alpha-helix_at_the_window_position_of_-2_(Qian-Sejnowski,_1988)

263	Weights_for_alpha-helix_at_the_window_position_of_-1_(Qian-Sejnowski,_1988)

264	Weights_for_alpha-helix_at_the_window_position_of_0_(Qian-Sejnowski,_1988)

265	Weights_for_alpha-helix_at_the_window_position_of_1_(Qian-Sejnowski,_1988)

266	Weights_for_alpha-helix_at_the_window_position_of_2_(Qian-Sejnowski,_1988)

267	Weights_for_alpha-helix_at_the_window_position_of_3_(Qian-Sejnowski,_1988)

268	Weights_for_alpha-helix_at_the_window_position_of_4_(Qian-Sejnowski,_1988)

269	Weights_for_alpha-helix_at_the_window_position_of_5_(Qian-Sejnowski,_1988)

270	Weights_for_alpha-helix_at_the_window_position_of_6_(Qian-Sejnowski,_1988)

271	Weights_for_beta-sheet_at_the_window_position_of_-6_(Qian-Sejnowski,_1988)

272	Weights_for_beta-sheet_at_the_window_position_of_-5_(Qian-Sejnowski,_1988)

273	Weights_for_beta-sheet_at_the_window_position_of_-4_(Qian-Sejnowski,_1988)

274	Weights_for_beta-sheet_at_the_window_position_of_-3_(Qian-Sejnowski,_1988)

275	Weights_for_beta-sheet_at_the_window_position_of_-2_(Qian-Sejnowski,_1988)

276	Weights_for_beta-sheet_at_the_window_position_of_-1_(Qian-Sejnowski,_1988)

277	Weights_for_beta-sheet_at_the_window_position_of_0_(Qian-Sejnowski,_1988)

278	Weights_for_beta-sheet_at_the_window_position_of_1_(Qian-Sejnowski,_1988)

279	Weights_for_beta-sheet_at_the_window_position_of_2_(Qian-Sejnowski,_1988)

280	Weights_for_beta-sheet_at_the_window_position_of_3_(Qian-Sejnowski,_1988)

281	Weights_for_beta-sheet_at_the_window_position_of_4_(Qian-Sejnowski,_1988)

282	Weights_for_beta-sheet_at_the_window_position_of_5_(Qian-Sejnowski,_1988)

283	Weights_for_beta-sheet_at_the_window_position_of_6_(Qian-Sejnowski,_1988)

284	Weights_for_coil_at_the_window_position_of_-6_(Qian-Sejnowski,_1988)

285	Weights_for_coil_at_the_window_position_of_-5_(Qian-Sejnowski,_1988)

286	Weights_for_coil_at_the_window_position_of_-4_(Qian-Sejnowski,_1988)

287	Weights_for_coil_at_the_window_position_of_-3_(Qian-Sejnowski,_1988)

288	Weights_for_coil_at_the_window_position_of_-2_(Qian-Sejnowski,_1988)

289	Weights_for_coil_at_the_window_position_of_-1_(Qian-Sejnowski,_1988)

290	Weights_for_coil_at_the_window_position_of_0_(Qian-Sejnowski,_1988)

291	Weights_for_coil_at_the_window_position_of_1_(Qian-Sejnowski,_1988)

292	Weights_for_coil_at_the_window_position_of_2_(Qian-Sejnowski,_1988)

293	Weights_for_coil_at_the_window_position_of_3_(Qian-Sejnowski,_1988)

294	Weights_for_coil_at_the_window_position_of_4_(Qian-Sejnowski,_1988)

295	Weights_for_coil_at_the_window_position_of_5_(Qian-Sejnowski,_1988)

296	Weights_for_coil_at_the_window_position_of_6_(Qian-Sejnowski,_1988)

297	Average_reduced_distance_for_C-alpha_(Rackovsky-Scheraga,_1977)

298	Average_reduced_distance_for_side_chain_(Rackovsky-Scheraga,_1977)

299	Side_chain_orientational_preference_(Rackovsky-Scheraga,_1977)

300	Average_relative_fractional_occurrence_in_A0(i)_(Rackovsky-Scheraga,_1982)

301	Average_relative_fractional_occurrence_in_AR(i)_(Rackovsky-Scheraga,_1982)

302	Average_relative_fractional_occurrence_in_AL(i)_(Rackovsky-Scheraga,_1982)

303	Average_relative_fractional_occurrence_in_EL(i)_(Rackovsky-Scheraga,_1982)

304	Average_relative_fractional_occurrence_in_E0(i)_(Rackovsky-Scheraga,_1982)

305	Average_relative_fractional_occurrence_in_ER(i)_(Rackovsky-Scheraga,_1982)

306	Average_relative_fractional_occurrence_in_A0(i-1)_(Rackovsky-Scheraga,_1982)

307	Average_relative_fractional_occurrence_in_AR(i-1)_(Rackovsky-Scheraga,_1982)

308	Average_relative_fractional_occurrence_in_AL(i-1)_(Rackovsky-Scheraga,_1982)

309	Average_relative_fractional_occurrence_in_EL(i-1)_(Rackovsky-Scheraga,_1982)

310	Average_relative_fractional_occurrence_in_E0(i-1)_(Rackovsky-Scheraga,_1982)

311	Average_relative_fractional_occurrence_in_ER(i-1)_(Rackovsky-Scheraga,_1982)

312	Value_of_theta(i)_(Rackovsky-Scheraga,_1982)

313	Value_of_theta(i-1)_(Rackovsky-Scheraga,_1982)

314	Transfer_free_energy_from_chx_to_wat_(Radzicka-Wolfenden,_1988)

315	Transfer_free_energy_from_oct_to_wat_(Radzicka-Wolfenden,_1988)

316	Transfer_free_energy_from_vap_to_chx_(Radzicka-Wolfenden,_1988)

317	Transfer_free_energy_from_chx_to_oct_(Radzicka-Wolfenden,_1988)

318	Transfer_free_energy_from_vap_to_oct_(Radzicka-Wolfenden,_1988)

319	Accessible_surface_area_(Radzicka-Wolfenden,_1988)

320	Energy_transfer_from_out_to_in(95%buried)_(Radzicka-Wolfenden,_1988)

321	Mean_polarity_(Radzicka-Wolfenden,_1988)

322	Relative_preference_value_at_N"_(Richardson-Richardson,_1988)

323	Relative_preference_value_at_N'_(Richardson-Richardson,_1988)

324	Relative_preference_value_at_N-cap_(Richardson-Richardson,_1988)

325	Relative_preference_value_at_N1_(Richardson-Richardson,_1988)

326	Relative_preference_value_at_N2_(Richardson-Richardson,_1988)

327	Relative_preference_value_at_N3_(Richardson-Richardson,_1988)

328	Relative_preference_value_at_N4_(Richardson-Richardson,_1988)

329	Relative_preference_value_at_N5_(Richardson-Richardson,_1988)

330	Relative_preference_value_at_Mid_(Richardson-Richardson,_1988)

331	Relative_preference_value_at_C5_(Richardson-Richardson,_1988)

332	Relative_preference_value_at_C4_(Richardson-Richardson,_1988)

333	Relative_preference_value_at_C3_(Richardson-Richardson,_1988)

334	Relative_preference_value_at_C2_(Richardson-Richardson,_1988)

335	Relative_preference_value_at_C1_(Richardson-Richardson,_1988)

336	Relative_preference_value_at_C-cap_(Richardson-Richardson,_1988)

337	Relative_preference_value_at_C'_(Richardson-Richardson,_1988)

338	Relative_preference_value_at_C"_(Richardson-Richardson,_1988)

339	Information_measure_for_alpha-helix_(Robson-Suzuki,_1976)

340	Information_measure_for_N-terminal_helix_(Robson-Suzuki,_1976)

341	Information_measure_for_middle_helix_(Robson-Suzuki,_1976)

342	Information_measure_for_C-terminal_helix_(Robson-Suzuki,_1976)

343	Information_measure_for_extended_(Robson-Suzuki,_1976)

344	Information_measure_for_pleated-sheet_(Robson-Suzuki,_1976)

345	Information_measure_for_extended_without_H-bond_(Robson-Suzuki,_1976)

346	Information_measure_for_turn_(Robson-Suzuki,_1976)

347	Information_measure_for_N-terminal_turn_(Robson-Suzuki,_1976)

348	Information_measure_for_middle_turn_(Robson-Suzuki,_1976)

349	Information_measure_for_C-terminal_turn_(Robson-Suzuki,_1976)

350	Information_measure_for_coil_(Robson-Suzuki,_1976)

351	Information_measure_for_loop_(Robson-Suzuki,_1976)

352	Hydration_free_energy_(Robson-Osguthorpe,_1979)

353	Mean_area_buried_on_transfer_(Rose_et_al.,_1985)

354	Mean_fractional_area_loss_(Rose_et_al.,_1985)

355	Side_chain_hydropathy,_uncorrected_for_solvation_(Roseman,_1988)

356	Side_chain_hydropathy,_corrected_for_solvation_(Roseman,_1988)

357	Loss_of_Side_chain_hydropathy_by_helix_formation_(Roseman,_1988)

358	Transfer_free_energy_(Simon,_1976),_Cited_by_Charton-Charton_(1982)

359	Principal_component_I_(Sneath,_1966)

360	Principal_component_II_(Sneath,_1966)

361	Principal_component_III_(Sneath,_1966)

362	Principal_component_IV_(Sneath,_1966)

363	Zimm-Bragg_parameter_s_at_20_C_(Sueki_et_al.,_1984)

364	Zimm-Bragg_parameter_sigma_x_1.0E4_(Sueki_et_al.,_1984)

365	Optimal_matching_hydrophobicity_(Sweet-Eisenberg,_1983)

366	Normalized_frequency_of_alpha-helix_(Tanaka-Scheraga,_1977)

367	Normalized_frequency_of_isolated_helix_(Tanaka-Scheraga,_1977)

368	Normalized_frequency_of_extended_structure_(Tanaka-Scheraga,_1977)

369	Normalized_frequency_of_chain_reversal_R_(Tanaka-Scheraga,_1977)

370	Normalized_frequency_of_chain_reversal_S_(Tanaka-Scheraga,_1977)

371	Normalized_frequency_of_chain_reversal_D_(Tanaka-Scheraga,_1977)

372	Normalized_frequency_of_left-handed_helix_(Tanaka-Scheraga,_1977)

373	Normalized_frequency_of_zeta_R_(Tanaka-Scheraga,_1977)

374	Normalized_frequency_of_coil_(Tanaka-Scheraga,_1977)

375	Normalized_frequency_of_chain_reversal_(Tanaka-Scheraga,_1977)

376	Relative_population_of_conformational_state_A_(Vasquez_et_al.,_1983)

377	Relative_population_of_conformational_state_C_(Vasquez_et_al.,_1983)

378	Relative_population_of_conformational_state_E_(Vasquez_et_al.,_1983)

379	Electron-ion_interaction_potential_(Veljkovic_et_al.,_1985)

380	Bitterness_(Venanzi,_1984)

381	Transfer_free_energy_to_lipophilic_phase_(von_Heijne-Blomberg,_1979)

382	Average_interactions_per_side_chain_atom_(Warme-Morgan,_1978)

383	RF_value_in_high_salt_chromatography_(Weber-Lacey,_1978)

384	Propensity_to_be_buried_inside_(Wertz-Scheraga,_1978)

385	Free_energy_change_of_epsilon(i)_to_epsilon(ex)_(Wertz-Scheraga,_1978)

386	Free_energy_change_of_alpha(Ri)_to_alpha(Rh)_(Wertz-Scheraga,_1978)

387	Free_energy_change_of_epsilon(i)_to_alpha(Rh)_(Wertz-Scheraga,_1978)

388	Polar_requirement_(Woese,_1973)

389	Hydration_potential_(Wolfenden_et_al.,_1981)

390	Principal_property_value_z1_(Wold_et_al.,_1987)

391	Principal_property_value_z2_(Wold_et_al.,_1987)

392	Principal_property_value_z3_(Wold_et_al.,_1987)

393	Unfolding_Gibbs_energy_in_water,_pH7.0_(Yutani_et_al.,_1987)

394	Unfolding_Gibbs_energy_in_water,_pH9.0_(Yutani_et_al.,_1987)

395	Activation_Gibbs_energy_of_unfolding,_pH7.0_(Yutani_et_al.,_1987)

396	Activation_Gibbs_energy_of_unfolding,_pH9.0_(Yutani_et_al.,_1987)

397	Dependence_of_partition_coefficient_on_ionic_strength_(Zaslavsky_et_al.,_1982)

398	Hydrophobicity_(Zimmerman_et_al.,_1968)

399	Bulkiness_(Zimmerman_et_al.,_1968)

400	Polarity_(Zimmerman_et_al.,_1968)

401	Isoelectric_point_(Zimmerman_et_al.,_1968)

402	RF_rank_(Zimmerman_et_al.,_1968)

403	Normalized_positional_residue_frequency_at_helix_termini_N4'(Aurora-Rose,_1998)

404	Normalized_positional_residue_frequency_at_helix_termini_N"'_(Aurora-Rose,_1998)

405	Normalized_positional_residue_frequency_at_helix_termini_N"_(Aurora-Rose,_1998)

406	Normalized_positional_residue_frequency_at_helix_termini_N'(Aurora-Rose,_1998)

407	Normalized_positional_residue_frequency_at_helix_termini_Nc_(Aurora-Rose,_1998)

408	Normalized_positional_residue_frequency_at_helix_termini_N1_(Aurora-Rose,_1998)

409	Normalized_positional_residue_frequency_at_helix_termini_N2_(Aurora-Rose,_1998)

410	Normalized_positional_residue_frequency_at_helix_termini_N3_(Aurora-Rose,_1998)

411	Normalized_positional_residue_frequency_at_helix_termini_N4_(Aurora-Rose,_1998)

412	Normalized_positional_residue_frequency_at_helix_termini_N5_(Aurora-Rose,_1998)

413	Normalized_positional_residue_frequency_at_helix_termini_C5_(Aurora-Rose,_1998)

414	Normalized_positional_residue_frequency_at_helix_termini_C4_(Aurora-Rose,_1998)

415	Normalized_positional_residue_frequency_at_helix_termini_C3_(Aurora-Rose,_1998)

416	Normalized_positional_residue_frequency_at_helix_termini_C2_(Aurora-Rose,_1998)

417	Normalized_positional_residue_frequency_at_helix_termini_C1_(Aurora-Rose,_1998)

418	Normalized_positional_residue_frequency_at_helix_termini_Cc_(Aurora-Rose,_1998)

419	Normalized_positional_residue_frequency_at_helix_termini_C'_(Aurora-Rose,_1998)

420	Normalized_positional_residue_frequency_at_helix_termini_C"_(Aurora-Rose,_1998)

421	Normalized_positional_residue_frequency_at_helix_termini_C"'_(Aurora-Rose,_1998)

422	Normalized_positional_residue_frequency_at_helix_termini_C4'_(Aurora-Rose,_1998)

423	Delta_G_values_for_the_peptides_extrapolated_to_0_M_urea_(O'Neil-DeGrado,_1990)

424	Helix_formation_parameters_(delta_delta_G)_(O'Neil-DeGrado,_1990)

425	Normalized_flexibility_parameters_(B-values),_average_(Vihinen_et_al.,_1994)

426	Normalized_flexibility_parameters_(B-values)_for_each_residue_surrounded_by_none_rigid_neighbours_(Vihinen_et_al.,_1994)

427	Normalized_flexibility_parameters_(B-values)_for_each_residue_surrounded_by_one_rigid_neighbours_(Vihinen_et_al.,_1994)

428	Normalized_flexibility_parameters_(B-values)_for_each_residue_surrounded_by_two_rigid_neighbours_(Vihinen_et_al.,_1994)

429	Free_energy_in_alpha-helical_conformation_(Munoz-Serrano,_1994)

430	Free_energy_in_alpha-helical_region_(Munoz-Serrano,_1994)

431	Free_energy_in_beta-strand_conformation_(Munoz-Serrano,_1994)

432	Free_energy_in_beta-strand_region_(Munoz-Serrano,_1994)

433	Free_energy_in_beta-strand_region_(Munoz-Serrano,_1994)

434	Free_energies_of_transfer_of_AcWl-X-LL_peptides_from_bilayer_interface_to_water_(Wimley-White,_1996)

435	Thermodynamic_beta_sheet_propensity_(Kim-Berg,_1993)

436	Turn_propensity_scale_for_transmembrane_helices_(Monne_et_al.,_1999)

437	Alpha_helix_propensity_of_position_44_in_T4_lysozyme_(Blaber_et_al.,_1993)

438	p-Values_of_mesophilic_proteins_based_on_the_distributions_of_B_values_(Parthasarathy-Murthy,_2000)

439	p-Values_of_thermophilic_proteins_based_on_the_distributions_of_B_values_(Parthasarathy-Murthy,_2000)

440	Distribution_of_amino_acid_residues_in_the_18_non-redundant_families_of_thermophilic_proteins_(Kumar_et_al.,_2000)

441	Distribution_of_amino_acid_residues_in_the_18_non-redundant_families_of_mesophilic_proteins_(Kumar_et_al.,_2000)

442	Distribution_of_amino_acid_residues_in_the_alpha-helices_in_thermophilic_proteins_(Kumar_et_al.,_2000)

443	Distribution_of_amino_acid_residues_in_the_alpha-helices_in_mesophilic_proteins_(Kumar_et_al.,_2000)

444	Side-chain_contribution_to_protein_stability_(kJ/mol)_(Takano-Yutani,_2001)

445	Propensity_of_amino_acids_within_pi-helices_(Fodje-Al-Karadaghi,_2002)

446	Hydropathy_scale_based_on_self-information_values_in_the_two-state_model_(5%_accessibility)_(Naderi-Manesh_et_al.,_2001)

447	Hydropathy_scale_based_on_self-information_values_in_the_two-state_model_(9%_accessibility)_(Naderi-Manesh_et_al.,_2001)

448	Hydropathy_scale_based_on_self-information_values_in_the_two-state_model_(16%_accessibility)_(Naderi-Manesh_et_al.,_2001)

449	Hydropathy_scale_based_on_self-information_values_in_the_two-state_model_(20%_accessibility)_(Naderi-Manesh_et_al.,_2001)

450	Hydropathy_scale_based_on_self-information_values_in_the_two-state_model_(25%_accessibility)_(Naderi-Manesh_et_al.,_2001)

451	Hydropathy_scale_based_on_self-information_values_in_the_two-state_model_(36%_accessibility)_(Naderi-Manesh_et_al.,_2001)

452	Hydropathy_scale_based_on_self-information_values_in_the_two-state_model_(50%_accessibility)_(Naderi-Manesh_et_al.,_2001)

453	Averaged_turn_propensities_in_a_transmembrane_helix_(Monne_et_al.,_1999)

454	Alpha-helix_propensity_derived_from_designed_sequences_(Koehl-Levitt,_1999)

455	Beta-sheet_propensity_derived_from_designed_sequences_(Koehl-Levitt,_1999)

456	Composition_of_amino_acids_in_extracellular_proteins_(percent)_(Cedano_et_al.,_1997)

457	Composition_of_amino_acids_in_anchored_proteins_(percent)_(Cedano_et_al.,_1997)

458	Composition_of_amino_acids_in_membrane_proteins_(percent)_(Cedano_et_al.,_1997)

459	Composition_of_amino_acids_in_intracellular_proteins_(percent)_(Cedano_et_al.,_1997)

460	Composition_of_amino_acids_in_nuclear_proteins_(percent)_(Cedano_et_al.,_1997)

461	Surface_composition_of_amino_acids_in_intracellular_proteins_of_thermophiles_(percent)_(Fukuchi-Nishikawa,_2001)

462	Surface_composition_of_amino_acids_in_intracellular_proteins_of_mesophiles_(percent)_(Fukuchi-Nishikawa,_2001)

463	Surface_composition_of_amino_acids_in_extracellular_proteins_of_mesophiles_(percent)_(Fukuchi-Nishikawa,_2001)

464	Surface_composition_of_amino_acids_in_nuclear_proteins_(percent)_(Fukuchi-Nishikawa,_2001)

465	Interior_composition_of_amino_acids_in_intracellular_proteins_of_thermophiles_(percent)_(Fukuchi-Nishikawa,_2001)

466	Interior_composition_of_amino_acids_in_intracellular_proteins_of_mesophiles_(percent)_(Fukuchi-Nishikawa,_2001)

467	Interior_composition_of_amino_acids_in_extracellular_proteins_of_mesophiles_(percent)_(Fukuchi-Nishikawa,_2001)

468	Interior_composition_of_amino_acids_in_nuclear_proteins_(percent)_(Fukuchi-Nishikawa,_2001)

469	Entire_chain_composition_of_amino_acids_in_intracellular_proteins_of_thermophiles_(percent)_(Fukuchi-Nishikawa,_2001)

470	Entire_chain_composition_of_amino_acids_in_intracellular_proteins_of_mesophiles_(percent)_(Fukuchi-Nishikawa,_2001)

471	Entire_chain_composition_of_amino_acids_in_extracellular_proteins_of_mesophiles_(percent)_(Fukuchi-Nishikawa,_2001)

472	Entire_chain_compositino_of_amino_acids_in_nuclear_proteins_(percent)_(Fukuchi-Nishikawa,_2001)

473	Amphiphilicity_index_(Mitaku_et_al.,_2002)

474	Volumes_including_the_crystallographic_waters_using_the_ProtOr_(Tsai_et_al.,_1999)

475	Volumes_not_including_the_crystallographic_waters_using_the_ProtOr_(Tsai_et_al.,_1999)

476	Electron-ion_interaction_potential_values_(Cosic,_1994)

477	Hydrophobicity_scales_(Ponnuswamy,_1993)

478	Hydrophobicity_coefficient_in_RP-HPLC,_C18_with_0.1%TFA/MeCN/H2O_(Wilce_et_al._1995)

479	Hydrophobicity_coefficient_in_RP-HPLC,_C8_with_0.1%TFA/MeCN/H2O_(Wilce_et_al._1995)

480	Hydrophobicity_coefficient_in_RP-HPLC,_C4_with_0.1%TFA/MeCN/H2O_(Wilce_et_al._1995)

481	Hydrophobicity_coefficient_in_RP-HPLC,_C18_with_0.1%TFA/2-PrOH/MeCN/H2O_(Wilce_et_al._1995)

482	Hydrophilicity_scale_(Kuhn_et_al.,_1995)

483	Retention_coefficient_at_pH_2_(Guo_et_al.,_1986)

484	Modified_Kyte-Doolittle_hydrophobicity_scale_(Juretic_et_al.,_1998)

485	Interactivity_scale_obtained_from_the_contact_matrix_(Bastolla_et_al.,_2005)

486	Interactivity_scale_obtained_by_maximizing_the_mean_of_correlation_coefficient_over_single-domain_globular_proteins_(Bastolla_et_al.,_2005)

487	Interactivity_scale_obtained_by_maximizing_the_mean_of_correlation_coefficient_over_pairs_of_sequences_sharing_the_TIM_barrel_fold_(Bastolla_et_al.,_2005)

488	Linker_propensity_index_(Suyama-Ohara,_2003)

489	Knowledge-based_membrane-propensity_scale_from_1D_Helix_in_MPtopo_databases_(Punta-Maritan,_2003)

490	Knowledge-based_membrane-propensity_scale_from_3D_Helix_in_MPtopo_databases_(Punta-Maritan,_2003)

491	Linker_propensity_from_all_dataset_(George-Heringa,_2003)

492	Linker_propensity_from_1-linker_dataset_(George-Heringa,_2003)

493	Linker_propensity_from_2-linker_dataset_(George-Heringa,_2003)

494	Linker_propensity_from_3-linker_dataset_(George-Heringa,_2003)

495	Linker_propensity_from_small_dataset_(linker_length_is_less_than_six_residues)_(George-Heringa,_2003)

496	Linker_propensity_from_medium_dataset_(linker_length_is_between_six_and_14

497	Linker_propensity_from_long_dataset_(linker_length_is_greater_than_14

498	Linker_propensity_from_helical_(annotated_by_DSSP)_dataset_(George-Heringa,_2003)

499	Linker_propensity_from_non-helical_(annotated_by_DSSP)_dataset_(George-Heringa,_2003)

500	The_stability_scale_from_the_knowledge-based_atom-atom_potential_(Zhou-Zhou,_2004)

501	The_relative_stability_scale_extracted_from_mutation_experiments_(Zhou-Zhou,_2004)

502	Buriability_(Zhou-Zhou,_2004)

503	Linker_index_(Bae_et_al.,_2005)

504	Mean_volumes_of_residues_buried_in_protein_interiors_(Harpaz_et_al.,_1994)

505	Average_volumes_of_residues_(Pontius_et_al.,_1996)

506	Hydrostatic_pressure_asymmetry_index,_PAI_(Di_Giulio,_2005)

507	Hydrophobicity_index_(Wolfenden_et_al.,_1979)

508	Average_internal_preferences_(Olsen,_1980)

509	Hydrophobicity-related_index_(Kidera_et_al.,_1985)

510	Apparent_partition_energies_calculated_from_Wertz-Scheraga_index_(Guy,_1985)

511	Apparent_partition_energies_calculated_from_Janin_index_(Guy,_1985)

512	Apparent_partition_energies_calculated_from_Chothia_index_(Guy,_1985)

513	Weights_from_the_IFH_scale_(Jacobs-White,_1989)

514	Hydrophobicity_index,_3.0_pH_(Cowan-Whittaker,_1990)

515	Scaled_side_chain_hydrophobicity_values_(Black-Mould,_1991)

516	Hydrophobicity_scale_from_native_protein_structures_(Casari-Sippl,_1992)

517	NNEIG_index_(Cornette_et_al.,_1987)

518	SWEIG_index_(Cornette_et_al.,_1987)

519	PRIFT_index_(Cornette_et_al.,_1987)

520	PRILS_index_(Cornette_et_al.,_1987)

521	ALTFT_index_(Cornette_et_al.,_1987)

522	ALTLS_index_(Cornette_et_al.,_1987)

523	TOTFT_index_(Cornette_et_al.,_1987)

524	TOTLS_index_(Cornette_et_al.,_1987)

525	Relative_partition_energies_derived_by_the_Bethe_approximation_(Miyazawa-Jernigan,_1999)

526	Optimized_relative_partition_energies_-_method_A_(Miyazawa-Jernigan,_1999)

527	Optimized_relative_partition_energies_-_method_B_(Miyazawa-Jernigan,_1999)

528	Optimized_relative_partition_energies_-_method_C_(Miyazawa-Jernigan,_1999)

529	Optimized_relative_partition_energies_-_method_D_(Miyazawa-Jernigan,_1999)

530	Hydrophobicity_index_(Engelman_et_al.,_1986)

531	Hydrophobicity_index_(Fasman,_1989)

532	GA_APV_OUT_(Heider,_Nov_2010)
}
\source{
}
\references{
Kawashima, S.; Ogata, H.; Kanehisa, M. Nucleic Acids Res 1999, 27, 368.
}
\examples{
}
\keyword{AAindex}
