\name{V3}
\alias{V3}
\docType{data}
\title{
HIV-1 V3 loop dataset for tropism prediction
}
\description{
HIV-1 V3 loop dataset containing V3 sequences of 200 X4 and 1151 R5 viruses.
}
\usage{data(V3)}
\format{
}
\details{
}
\source{
}
\references{
Dybowski J.N., Heider D., Hoffmann D. PLoS Computational Biology 2010, 6(4): e1000743.
}
\examples{
}
\keyword{HIV}
